Programming exercises: Fix missing gradle feedback on Jenkins (#6802)
